feat: create method that makes insertBreak default element
