+++ packages/element-heading/src/level01/index.ts
@@ -1,4 +1,4 @@
-import { Element } from '@bento-editor/core';
+import { Element, helpers } from '@bento-editor/core';
 import attributes, { Attributes } from './attributes';
 import editable from './editable';
 import toolbox from './toolbox';
@@ -9,5 +9,7 @@ const element: Element<Attributes> = {
   attributes,
   editable,
   toolbox,
+  insertBreak: (editor, nodeEntry, config) =>
+    helpers.insertBreaks.defaultElementInsertBreak(editor, nodeEntry, config),
 };
 export default element;
